Remove background
The background had to be removed for two reasons. First of all, due to
the "position: fixed;" tag in the css, it was scrolling oddly on mobile
phones. Additionally, it was rendering over the background color without
the alpha, likely due to the fact that image was converted from a
JPEG-2000 file to a JPEG file, which will be fixed in a later update.